Make vid.width/height unsigned and clean up the mess.

+++ libs/video/renderer/sw/draw.c
@@ -553,7 +553,7 @@ Draw_SubPic (int x, int y, qpic_t *pic, int srcx, int srcy, int width,
 		width += x;
 		x = 0;
 	}
-	if (x + width > vid.width)
+	if ((unsigned) (x + width) > vid.width)
 		width = vid.width - x;
 	if (width <= 0)
 		return;
@@ -562,7 +562,7 @@ Draw_SubPic (int x, int y, qpic_t *pic, int srcx, int srcy, int width,
 		height += y;
 		y = 0;
 	}
-	if (y + height > vid.height)
+	if ((unsigned) (y + height) > vid.height)
 		height = vid.height - y;
 	if (height <= 0)
 		return;
@@ -696,7 +696,7 @@ Draw_TileClear (int x, int y, int w, int h)
 	byte       *psrc;
 	vrect_t     vr;
 
-	CLIP (x, y, w, h, vid.width, vid.height);
+	CLIP (x, y, w, h, (int) vid.width, (int) vid.height);
 
 	r_rectdesc.rect.x = x;
 	r_rectdesc.rect.y = y;
@@ -764,7 +764,7 @@ Draw_Fill (int x, int y, int w, int h, int c)
 		Sys_MaskPrintf (SYS_VID, "Bad Draw_Fill(%d, %d, %d, %d, %c)\n",
 						x, y, w, h, c);
 	}
-	CLIP (x, y, w, h, vid.width, vid.height);
+	CLIP (x, y, w, h, (int) vid.width, (int) vid.height);
 
 	dest = ((byte*)vid.buffer) + y * vid.rowbytes + x;
 	for (v = 0; v < h; v++, dest += vid.rowbytes)
